jatin/requirements.txt
gradio
spacy
pdfplumber
python-docx
fuzzywuzzy
